refactor(api): set flowVersion.updatedBy to null if user is deleted
