+++ packages/server/api/src/app/database/postgres-connection.ts
@@ -113,7 +113,8 @@ import { DropUnusedPlatformIndex1709500873378 } from './migration/postgres/17095
 import { SetNotNullOnPlatform1709505632771 } from './migration/1709505632771-SetNotNullOnPlatform'
 import { AddPlatformForeignKeyToProjectPostgres1709566642531 } from './migration/postgres/1709566642531-add-platform-foreign-key-to-project-postgres'
 import { AddUserEmailToReferral1709500213947 } from './migration/postgres/1709500213947-add-user-email-to-referral'
-import { AddGeneratedByEmailToSigningKeyPostgres1709638978925 } from './migration/1709638978925-add-generated-by-email-to-signing-key-postgres'
+import { AddGeneratedByEmailToSigningKeyPostgres1709638978925 } from './migration/postgres/1709638978925-add-generated-by-email-to-signing-key-postgres'
+import { SetFlowVersionUpdatedByToNullIfUserIsDeletedPostgres1709641016072 } from './migration/postgres/1709641016072-set-flow-version-updated-by-to-null-if-user-is-deleted-postgres'
 
 const getSslConfig = (): boolean | TlsOptions => {
     const useSsl = system.get(SystemProp.POSTGRES_USE_SSL)
@@ -186,6 +187,7 @@ const getMigrations = (): (new () => MigrationInterface)[] => {
         AddTriggerTestStrategy1707087022764,
         AddCategoriesToPieceMetadataPostgres1707231704973,
         AddUniqueStoreConstraint1708521505204,
+        SetFlowVersionUpdatedByToNullIfUserIsDeletedPostgres1709641016072,
     ]
 
     const edition = getEdition()
