+++ packages/server/api/src/app/flows/flow-version/flow-version-entity.ts
@@ -46,7 +46,7 @@ export const FlowVersionEntity = new EntitySchema<FlowVersionSchema>({
             type: 'many-to-one',
             target: 'user',
             cascade: true,
-            onDelete: 'CASCADE',
+            onDelete: 'SET NULL',
             joinColumn: {
                 name: 'updatedBy',
                 foreignKeyConstraintName: 'fk_updated_by_user_flow',
